+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Task Lists are now vertically aligned with their counterparts.
 - If the selected Document shows up in Search Results, it's now highlighted. Same thing when the search is closed. Apparently I fixed that at some point. Cool.
 - The Search is now case-insensitive.
+- Image Drag & Drop now works.
 
 ### Notes